+++ src/app/(main)/_components/molecules/SelectAgoraColor.tsx
@@ -3,18 +3,10 @@
 import React, { useState } from 'react';
 import { useCreateAgora } from '@/store/create';
 import { useShallow } from 'zustand/react/shallow';
+import COLOR from '@/constants/agoraColor';
 import AgoraPointColor from '../atoms/AgoraPointColor';
 
-const COLOR = [
-  { value: 'bg-agora-point-color1', label: '청록색' },
-  { value: 'bg-agora-point-color2', label: '밝은 민트색' },
-  { value: 'bg-agora-point-color3', label: '연한 갈색' },
-  { value: 'bg-agora-point-color4', label: '연한 주황색' },
-  { value: 'bg-agora-point-color5', label: '연한 핑크색' },
-  { value: 'bg-agora-point-color6', label: '남색' },
-];
-
-function AgoraPointColorList() {
+export default function SelectAgoraColor() {
   const [isCheck, setIsCheck] = useState<number>(0);
   const { setColor } = useCreateAgora(
     useShallow((state) => ({ setColor: state.setColor })),
@@ -44,5 +36,3 @@ function AgoraPointColorList() {
     </div>
   );
 }
-
-export default React.memo(AgoraPointColorList);
